Fixed my mess
botutils.py
Lines 79-84 & 133-134

GodManager.py
Lines 171-178

AdminManager.py
Lines 147-154

BotManager.py
Lines 49-51

database.py
Lines 199-202
