+++ cogs/BotManager.py
@@ -46,6 +46,9 @@ class BotManager(commands.Cog, name="Bot Commands"):
                                "\n"
                                "  <> = required argument. [] = optional argument.\n"
                                "\n"
+                               "> Change the name of your God:\n"
+                               "    " + prefix + " rename <new name>          "
+                                                 "Changes God's name.\n"
                                "> Set your religion to invite-only:\n"
                                "    " + prefix + "access                      "
                                                  "Set your religion as open or invite only.\n"
